Move Copyright component to common directory
