+++ app/components/common/Copyright/Copyright.tsx
