+++ app/components/common/Copyright/index.ts
