+++ app/tag/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import { Copyright } from "../../components/concerns/Copyright";
+import { Copyright } from "../../components/common/Copyright";
 import { EntryList } from "../../components/concerns/EntryList";
 import { WithSiteTitle } from "../../constants";
 import { metadata } from "../../layout";
